file structure refactor; restructure for src/data_input and src/data_analysis
