+++ src/data_analysis/mod.rs
@@ -1,0 +1,6 @@
+// src/data_analysis/mod.rs
+
+pub mod calc_step_response;
+pub mod fft_utils;
+
+// src/data_analysis/mod.rs
