+++ src/data_input/mod.rs
@@ -1,0 +1,6 @@
+// src/data_input/mod.rs
+
+pub mod log_data;
+pub mod log_parser;
+
+// src/data_input/mod.rs
